rename OutputChannel to PulseChannel
