+++ qiskit/pulse/commands/persistent_value.py
@@ -9,7 +9,7 @@
 Persistent value.
 """
 
-from qiskit.pulse.device import OutputChannel
+from qiskit.pulse.device import PulseChannel
 from qiskit.pulse.timeslots import Interval, Timeslot, TimeslotCollection
 from qiskit.pulse.exceptions import PulseError
 from .instruction import Instruction
@@ -53,14 +53,14 @@ class PersistentValue(Command):
     def __repr__(self):
         return '%s(%s, value=%s)' % (self.__class__.__name__, self.name, self.value)
 
-    def __call__(self, channel: OutputChannel) -> 'PersistentValueInstruction':
+    def __call__(self, channel: PulseChannel) -> 'PersistentValueInstruction':
         return PersistentValueInstruction(self, channel)
 
 
 class PersistentValueInstruction(Instruction):
     """Instruction to keep persistent value. """
 
-    def __init__(self, command: PersistentValue, channel: OutputChannel, start_time: int = 0):
+    def __init__(self, command: PersistentValue, channel: PulseChannel, start_time: int = 0):
         slots = [Timeslot(Interval(start_time, start_time), channel)]
         super().__init__(command, start_time, TimeslotCollection(slots))
         self._channel = channel
@@ -71,8 +71,8 @@ class PersistentValueInstruction(Instruction):
         return self._command
 
     @property
-    def channel(self) -> OutputChannel:
-        """OutputChannel channel."""
+    def channel(self) -> PulseChannel:
+        """The channel to be instructed. """
         return self._channel
 
     def __repr__(self):
