+++ qiskit/pulse/device/__init__.py
@@ -8,8 +8,8 @@
 """Channel classes for pulse."""
 
 from .device_specification import DeviceSpecification
-from .output_channel import DriveChannel, ControlChannel, MeasureChannel
-from .output_channel import OutputChannel
+from .pulse_channels import DriveChannel, ControlChannel, MeasureChannel
+from .pulse_channels import PulseChannel
 from .channels import AcquireChannel, MemorySlot, RegisterSlot, SnapshotChannel
 from .channels import Channel
 from .qubit import Qubit
