+++ qiskit/pulse/device/device_specification.py
@@ -13,7 +13,7 @@ from typing import List
 
 from qiskit.pulse.exceptions import PulseError
 from qiskit.validation.exceptions import ModelValidationError
-from .output_channel import DriveChannel, ControlChannel, MeasureChannel
+from .pulse_channels import DriveChannel, ControlChannel, MeasureChannel
 from .channels import AcquireChannel, MemorySlot, RegisterSlot
 from .qubit import Qubit
