+++ qiskit/pulse/device/qubit.py
@@ -11,7 +11,7 @@ Physical qubit.
 from typing import List
 
 from qiskit.pulse.exceptions import PulseError
-from .output_channel import DriveChannel, ControlChannel, MeasureChannel
+from .pulse_channels import DriveChannel, ControlChannel, MeasureChannel
 from .channels import AcquireChannel
